Fix flaky test for n-of-n integer ElGamal
