+++ scicrypt-bigint/src/modular/pow.rs
@@ -58,3 +58,71 @@ impl BigInteger {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::BigInteger;
+
+    #[test]
+    fn test_powmod_small_base() {
+        let b = BigInteger::from_string("105".to_string(), 10, 7);
+        let e = BigInteger::from_string("92848022024833655041372304737256052921065477715975001419347548380734496823522565044177931242947122534563813415992433917108481569319894167972639736788613656007853719476736625612543893748136536594494005487213485785676333621181690463942417781763743640447405597892807333854156631166426238815716390011586838580891".to_string(), 10, 1024);
+        let m = BigInteger::from_string("149600854933825512159828331527177109689118555212385170831387365804008437367913613643959968668965614270559113472851544758183282789643129469226548555150464780229538086590498853718102052468519876788192865092229749643546710793464305243815836267024770081889047200172952438000587807986096107675012284269101785114471".to_string(), 10, 1024);
+
+        let res = b.pow_mod(&e, &m);
+
+        let expected = BigInteger::from_string("93381698043531945590460734835437626929406390544089092303961497613088223192062266567807404255983003371786424645697784253062005750244340967243067126193405796382070980127325598311265307429963380264226672935938163271489566200721235534991781171263956580735259196276780705026850011214281556290838394235159210861122".to_string(), 10, 1024);
+        assert_eq!(res, expected);
+    }
+
+    #[test]
+    fn test_powmod_small_base_oversized() {
+        let b = BigInteger::from_string("105".to_string(), 10, 1024);
+        let e = BigInteger::from_string("92848022024833655041372304737256052921065477715975001419347548380734496823522565044177931242947122534563813415992433917108481569319894167972639736788613656007853719476736625612543893748136536594494005487213485785676333621181690463942417781763743640447405597892807333854156631166426238815716390011586838580891".to_string(), 10, 1024);
+        let m = BigInteger::from_string("149600854933825512159828331527177109689118555212385170831387365804008437367913613643959968668965614270559113472851544758183282789643129469226548555150464780229538086590498853718102052468519876788192865092229749643546710793464305243815836267024770081889047200172952438000587807986096107675012284269101785114471".to_string(), 10, 1024);
+
+        let res = b.pow_mod(&e, &m);
+
+        let expected = BigInteger::from_string("93381698043531945590460734835437626929406390544089092303961497613088223192062266567807404255983003371786424645697784253062005750244340967243067126193405796382070980127325598311265307429963380264226672935938163271489566200721235534991781171263956580735259196276780705026850011214281556290838394235159210861122".to_string(), 10, 1024);
+        assert_eq!(res, expected);
+    }
+
+    #[test]
+    fn test_powmod_small_exponent() {
+        let b = BigInteger::from_string("92848022024833655041372304737256052921065477715975001419347548380734496823522565044177931242947122534563813415992433917108481569319894167972639736788613656007853719476736625612543893748136536594494005487213485785676333621181690463942417781763743640447405597892807333854156631166426238815716390011586838580891".to_string(), 10, 1024);
+        let e = BigInteger::from_string("105".to_string(), 10, 7);
+        let m = BigInteger::from_string("149600854933825512159828331527177109689118555212385170831387365804008437367913613643959968668965614270559113472851544758183282789643129469226548555150464780229538086590498853718102052468519876788192865092229749643546710793464305243815836267024770081889047200172952438000587807986096107675012284269101785114471".to_string(), 10, 1024);
+
+        let res = b.pow_mod(&e, &m);
+
+        // TODO: Validate this number
+        let expected = BigInteger::from_string("75449268817968422679819900589734348654486644392551728445064418436053449491480437746932914650717830240874061893534937751643365068436165993034818308531811356620889371580247889632561792360083344802209721380578912179116118493677119654295291184624591629851342172735975592027041999972633543293770666292467255672690".to_string(), 10, 1024);
+        assert_eq!(res, expected);
+    }
+
+    #[test]
+    fn test_powmod_mini() {
+        let b = BigInteger::from(3u64);
+        let e = BigInteger::from(7u64);
+        let m = BigInteger::from(11u64);
+
+        let res = b.pow_mod(&e, &m);
+
+        // TODO: Validate this number
+        let expected = BigInteger::from_string("9".to_string(), 10, 1024);
+        assert_eq!(res, expected);
+    }
+
+    #[test]
+    fn test_powmod_mini_plusmod() {
+        let b = BigInteger::from(14u64);
+        let e = BigInteger::from(7u64);
+        let m = BigInteger::from(11u64);
+
+        let res = b.pow_mod(&e, &m);
+
+        // TODO: Validate this number
+        let expected = BigInteger::from_string("9".to_string(), 10, 1024);
+        assert_eq!(res, expected);
+    }
+}
